requirements.txt
streamlit>=1.24.0
tensorflow>=2.13.0
numpy>=1.24.3
Pillow>=9.5.0 